+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Web-based overlay with current song ([#22]).
 - Player will no longer pause the current song (if it's playing) and will instead synchronize the state of the player with Spotify ([#18]).
 - Implement `!commands rename <from> <to>`
+- Ability to sync remote state of player with `[player] sync_player_interval = "10s"` ([#18]).
 
 ### Changed
 - Cleaned up old cruft in the codebase (`gfx` module).